+++ Chronolibris.Domain/Entities/PersonRole.cs
@@ -8,7 +8,7 @@ namespace Chronolibris.Domain.Entities
 {
     public class PersonRole
     {
-        public required int Id { get; set; }
+        public required long Id { get; set; }
         public required string Name { get; set; }
     }
 }
