+++ Chronolibris.Domain/Entities/Publisher.cs
@@ -13,7 +13,7 @@ namespace Chronolibris.Domain.Entities
         public required string Description { get; set; }
         public required DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
-        public required int CountryId { get; set; }
+        public required long CountryId { get; set; }
         public ICollection<Series> Series { get; set; } = new List<Series>();
         public ICollection<Book> Books { get; set; } = new List<Book>();
     }
